moving ownership of selected channel
